backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.0
google-cloud-bigquery==3.14.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
python-dotenv==1.0.0
websockets==12.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-cloud-trace==1.6.0
opentelemetry-exporter-cloud-monitoring==1.6.0